backend/requirements.txt
# Core FastAPI dependencies
# FastAPI and server
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.9

# Authentication
PyJWT==2.10.1
python-jose==3.5.0

# Pydantic for data validation
pydantic==2.11.7
pydantic_core==2.33.2
email-validator==2.3.0

# Supabase client
supabase==2.17.0
postgrest==1.1.1
storage3==0.12.0
gotrue==2.12.3
realtime==2.6.0

# HTTP client for LinkedIn API
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.2.0

# OpenAI for content generation
openai==2.6.1

# Environment variables
python-dotenv==1.1.1
# Common dependencies
starlette==0.46.2
anyio==4.9.0
sniffio==1.3.1
idna==3.10
certifi==2025.7.14
typing-extensions==4.14.1

#QDrant for vectordb
qdrant-client==1.16.1
